HA update #14 & HACS updates #16

+++ custom_components/gtfs2/gtfs_helper.py
@@ -389,7 +389,7 @@ def get_gtfs(hass, path, data, update=False):
         _LOGGER.warning("Cannot use this datasource as still unpacking: %s", filename)
         return "extracting"
     if update and data["extract_from"] == "url" and os.path.exists(os.path.join(gtfs_dir, file)):
-        remove_datasource(hass, path, filename)
+        remove_datasource(hass, path, filename, False)
     if update and data["extract_from"] == "zip" and os.path.exists(os.path.join(gtfs_dir, file)) and os.path.exists(os.path.join(gtfs_dir, sqlite)):
         os.remove(os.path.join(gtfs_dir, sqlite))      
     if data["extract_from"] == "zip":
@@ -404,10 +404,15 @@ def get_gtfs(hass, path, data, update=False):
             except Exception as ex:  # pylint: disable=broad-except
                 _LOGGER.error("The given URL or GTFS data file/folder was not found")
                 return "no_data_file"                
-    (gtfs_root, _) = os.path.splitext(file)
-
+    
+    # if update (servicecall) then check if new file does not only have future dates
+    if update and not check_calendar_dates_from_zip(gtfs_dir, file):
+        _LOGGER.info('New file contains only dates in the future, extracting terminated')
+        return
+    
+    (gtfs_root, _) = os.path.splitext(file)    
     sqlite_file = f"{gtfs_root}.sqlite?check_same_thread=False"
-    joined_path = os.path.join(gtfs_dir, sqlite_file) 
+    joined_path = os.path.join(gtfs_dir, sqlite_file)     
     gtfs = pygtfs.Schedule(joined_path)
     if not gtfs.feeds: 
         if data.get("clean_feed_info", False):
@@ -428,6 +433,51 @@ def extract_from_zip(hass, gtfs, gtfs_dir, file, remove_file):
         return
     pygtfs.append_feed(gtfs, os.path.join(gtfs_dir, file))
     check_datasource_index(hass, gtfs, gtfs_dir, file[:-4])
+    
+def check_calendar_dates_from_zip(gtfs_dir,file):
+    _LOGGER.debug("Checking if file contains only future data: %s ", file)
+    filename = os.path.join(gtfs_dir, file)
+    # Rename existing sqlite if existing (i.e. in case of a fresh install)
+    if os.path.exists(os.path.join(gtfs_dir, file[:-4] + ".sqlite")):
+        os.rename (os.path.join(gtfs_dir, file[:-4] + '.sqlite'), os.path.join(gtfs_dir, file[:-4] + '.sqlite_current'))
+    #Load the ZIP archive
+    zin = zipfile.ZipFile (f"{os.path.join(gtfs_dir, filename)}", 'r')
+    check_list=[]
+    try:
+        for item in zin.infolist():
+            if item.filename[0:8] == 'calendar' :
+                if item.filename == 'calendar.txt':
+                    column = 'start_date'
+                else:
+                    column = 'date'
+                with open(zin.extract(item.filename)) as f:
+                    header = f.readline().strip('\n')   #
+                    data = f.readlines() 
+                    index =header.replace('"','').split(',').index(column)           
+                    list = []
+                    for line in data:
+                        list.append(line.split(',')[index])
+                    check_list.append(min(list))
+        min_date = datetime.datetime.strptime(min(check_list),"%Y%m%d")
+        _LOGGER.debug("Youngest calender date from new files: %s, is: %s", check_list, min_date)
+        if min_date > datetime.datetime.now()  :
+            _LOGGER.info("New file contains only dates in the future, keeping current")
+            if os.path.exists(os.path.join(gtfs_dir, file[:-4] + ".sqlite")):
+                os.remove(os.path.join(gtfs_dir, file[:-4] + ".sqlite"))
+            os.rename (os.path.join(gtfs_dir, file[:-4] + '.sqlite_current'), os.path.join(gtfs_dir, file[:-4] + '.sqlite'))
+            return False
+    except Exception as ex:
+        _LOGGER.error("Error getting earliest dates from zip, continuing with extract, error: %s", ex)
+        _LOGGER.debug(f"Removing/restoring sqlite after error")
+        if os.path.exists(os.path.join(gtfs_dir, file[:-4] + ".sqlite")):
+            os.remove(os.path.join(gtfs_dir, file[:-4] + ".sqlite"))
+        if os.path.exists(os.path.join(gtfs_dir, file[:-4] + ".sqlite")):
+            os.rename (os.path.join(gtfs_dir, file[:-4] + '.sqlite'), os.path.join(gtfs_dir, file[:-4] + '.sqlite_current'))
+        return False
+    _LOGGER.debug(f"New file is not containing only newer dates, removing current/copied sqlite")    
+    if os.path.exists(os.path.join(gtfs_dir, file[:-4] + ".sqlite_current")):
+        os.remove(os.path.join(gtfs_dir, file[:-4] + ".sqlite_current"))
+    return True
 
 def remove_from_zip(delmelist,gtfs_dir,file):
     _LOGGER.debug("Removing data: %s , from zipfile: %s", delmelist, file)
@@ -540,10 +590,10 @@ def get_datasources(hass, path) -> dict[str]:
     _LOGGER.debug(f"Datasources in folder: {datasources}")
     return datasources
 
-def remove_datasource(hass, path, filename):
+def remove_datasource(hass, path, filename, include_sqlite):
     gtfs_dir = hass.config.path(path)
     _LOGGER.info(f"Removing datasource: {os.path.join(gtfs_dir, filename)}.*")
-    if os.path.exists(os.path.join(gtfs_dir, filename + ".sqlite")):
+    if include_sqlite and os.path.exists(os.path.join(gtfs_dir, filename + ".sqlite")):
         os.remove(os.path.join(gtfs_dir, filename + ".sqlite"))
     if os.path.exists(os.path.join(gtfs_dir, filename + "_temp.zip")):     
         os.remove(os.path.join(gtfs_dir, filename + "_temp.zip"))
@@ -716,7 +766,7 @@ def get_local_stop_list(hass, schedule, data):
     device_tracker = hass.states.get(data['device_tracker_id'])
     latitude = device_tracker.attributes.get("latitude", None)
     longitude = device_tracker.attributes.get("longitude", None) 
-    radius = data.get("radius", DEFAULT_LOCAL_STOP_RADIUS) / 130000
+    radius = data.get("radius", DEFAULT_LOCAL_STOP_RADIUS) / 111111
     sql_query = f"""
         SELECT stop.stop_id, stop.stop_name
         FROM stops stop
@@ -765,8 +815,8 @@ def get_local_stops_next_departures(self):
     tomorrow_calendar_date_where = f"AND (calendar_date_today.date = date(:now_offset))"
     time_range = str('+' + str(self._data.get("timerange", DEFAULT_LOCAL_STOP_TIMERANGE)) + ' minute')
     time_range_history = str('-' + str(self._data.get("timerange_history", DEFAULT_LOCAL_STOP_TIMERANGE_HISTORY)) + ' minute')
-    radius = self._data.get("radius", DEFAULT_LOCAL_STOP_RADIUS) / 130000
-    if not latitude or not latitude:
+    radius = self._data.get("radius", DEFAULT_LOCAL_STOP_RADIUS) / 111111
+    if not latitude or not longitude:
         _LOGGER.error("No latitude and/or longitude for : %s", self._data['device_tracker_id'])
         return []
     if include_tomorrow:
@@ -935,4 +985,3 @@ async def update_gtfs_local_stops(hass, data):
         _LOGGER.debug("Reloading local stops for config_entry_id: %s", cf_entry) 
         reload = await hass.config_entries.async_reload(cf_entry)    
     return
-
